+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - Define attribute type classes as parameters
 - Products on which mass edit operation is not performed are also ignored from operation finalize method
 - Create specific serializer service for versioning
+- Create a metric factory
 
 ## Bug fixes
 - Replaced usage of Symfony process to launch background job with a simple exec, more reliable on a heavily loaded environment
@@ -73,6 +74,7 @@
 - Upload on a job with a custom step (non ItemStep) doesn't crash anymore
 - Memory leak fixed in pim:version:refresh command
 - Fixed a bug when try to remove the family of a product
+- Wrong date conversion fixes on grid and form
 
 ## BC breaks
 - Remove FlexibleEntityBundle